% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{Dire Bonjour}
\usage{
dire_bonjour(prenom = "toi")
}
\arguments{
\item{prenom}{character : Name of the person to greet. Default is "toi".}
}
\value{
character : A greeting message.
}
\description{
Dire Bonjour
}
\details{
This function takes a name as input and returns a greeting message in the format "Bonjour \{prenom\} !". If no name is provided, it defaults to "toi".
}
\examples{
dire_bonjour(prenom = "Vincent")
dire_bonjour()
}
