% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pays.R
\docType{data}
\name{pays}
\alias{pays}
\title{pays}
\format{
A data frame with 241 rows and 5 variables:
\describe{
\item{ ISO_3166_1_num }{  integer }
\item{ ISO_3166_1_alpha2 }{  character }
\item{ ISO_3166_1_alpha3 }{  character }
\item{ nom_francais }{  character }
\item{ nom_anglais }{  character }
}
}
\source{
Source
}
\usage{
pays
}
\description{
Liste des pays ainsi que des codes ISO 3166-1 associés.
}
\keyword{datasets}
